# core-lending-provisioning-risk-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-provisioning-risk
  description: Provisioning Risk Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/provisioning-cases/{provisioningCaseId}/calculations/{provisioningCalculationId}/journal/{provisioningJournalId}:
    get:
      tags:
        - ProvisioningJournal
      summary: Get a provisioning journal entry by ID
      operationId: getById_4
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningJournalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningJournalDTO'
    put:
      tags:
        - ProvisioningJournal
      summary: Update a provisioning journal entry
      operationId: update_4
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningJournalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningJournalDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningJournalDTO'
    delete:
      tags:
        - ProvisioningJournal
      summary: Delete a provisioning journal entry
      operationId: delete_4
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningJournalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/provisioning-cases/{provisioningCaseId}/calculations/{provisioningCalculationId}/journal:
    get:
      tags:
        - ProvisioningJournal
      summary: List or search provisioning journal entries for a calculation
      operationId: findAll_4
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.provisioningCalculationId
          in: query
          description: Exact filter for provisioning calculation id
          required: false
          schema:
            type: string
        - name: filters.accountingJournalEntryId
          in: query
          description: Exact filter for accounting journal entry id
          required: false
          schema:
            type: string
        - name: filters.provisionChangeAmount
          in: query
          description: Exact filter for provision change amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[provisionChangeAmount].from
          in: query
          description: Filter provision change amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[provisionChangeAmount].to
          in: query
          description: Filter provision change amount to value
          required: false
          schema:
            type: string
        - name: filters.postedAt
          in: query
          description: Exact filter for posted at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[postedAt].from
          in: query
          description: Filter posted at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[postedAt].to
          in: query
          description: Filter posted at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.postingDescription
          in: query
          description: Exact filter for posting description
          required: false
          schema:
            type: string
        - name: filters.isReversal
          in: query
          description: Exact filter for is reversal
          required: false
          schema:
            type: boolean
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseProvisioningJournalDTO'
    post:
      tags:
        - ProvisioningJournal
      summary: Create a new provisioning journal entry
      operationId: create_4
      parameters:
        - name: provisioningCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: provisioningCalculationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProvisioningJournalDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProvisioningJournalDTO'
components:
  schemas:
...
    ProvisioningJournalDTO:
      required:
        - accountingJournalEntryId
        - isReversal
        - postedAt
        - postingDescription
        - provisionChangeAmount
        - provisioningCalculationId
      type: object
      properties:
        provisioningJournalId:
          type: string
          format: uuid
          readOnly: true
        provisioningCalculationId:
          type: string
          format: uuid
        accountingJournalEntryId:
          type: string
          format: uuid
        provisionChangeAmount:
          type: number
        postedAt:
          type: string
          format: date-time
        postingDescription:
          maxLength: 500
          minLength: 0
          type: string
        isReversal:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseProvisioningJournalDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProvisioningJournalDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.